requirements.txt
asgiref==3.5.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-stripe==2.6.1
django==3.2.12
django-filter==21.1
django-leaflet==0.28.2
django-merchant==0.2
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.2
itypes==1.2.0
jinja2==3.0.3
jsonfield==3.1.0
libcxx==12.0.1
libffi==3.4.2
libzlib==1.2.11
markdown==3.3.6
markupsafe==2.1.1
ncurses==6.3
openapi-codec==1.3.2
openssl==3.0.0
pip==22.0.4
psycopg2-binary==2.9.3
pytz==2021.3
pyyaml==6.0
readline==8.1
requests==2.27.1
setuptools==60.9.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
tk==8.6.12
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wheel==0.37.1
xz==5.2.5
zipp==3.7.0
zlib==1.2.11
